requirements.txt
# Core ML frameworks
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Transformers and LLM tools
transformers>=4.40.0
accelerate>=0.27.0
peft>=0.10.0
bitsandbytes>=0.43.0
unsloth>=2024.4

# TRL for GRPO training
trl>=0.8.0

# Data handling
datasets>=2.18.0
pandas>=2.0.0
numpy>=1.24.0

# TensorFlow for reading AudioSet TFRecords
tensorflow>=2.15.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Experiment tracking (optional)
wandb>=0.16.0

# Evaluation metrics
nltk>=3.8.0
pycocoevalcap>=1.2

# Utilities
tqdm>=4.66.0
requests>=2.31.0
gdown>=5.1.0

# For downloading from Google Cloud Storage
google-cloud-storage>=2.14.0
